// src/components/RoundsToWinInput/RoundsToWinInput.types.ts
import type { HTMLAttributes, ReactNode } from "react";

export type RoundsToWinOption = {
  label: ReactNode;
  value: string;
  disabled?: boolean;
};

export type RoundsToWinInputProps = Omit<HTMLAttributes<HTMLDivElement>, "defaultValue" | "onChange"> & {
  label?: ReactNode;
  options?: RoundsToWinOption[];
  value?: string;
  defaultValue?: string;
  onChange?: (value: string, option: RoundsToWinOption) => void;
  disabled?: boolean;
  name?: string;
};